moved http get request in sitesettings.service.ts to prevent redundant requests. Added input checking. bundled the hidden field checks to prevent multiple subscriptions per component
