+++ src/main/webapp/app/shared/subject/subject-dialog.component.html
@@ -51,50 +51,50 @@
                 </option>
             </select>
         </div>
+        <div *ngIf="(siteSettings$ | async).hiddenSubjectFields as hiddenFields">
+            <div *ngIf="!hiddenFields.includes(HideableSubjectField.GROUP)" class="form-group">
+                <label class="form-control-label"
+                    for="field_group"
+                    [translate]="'managementPortalApp.subject.group'">
+                </label>
+                <select class="form-control"
+                    id="field_group"
+                    name="group"
+                    [(ngModel)]="groupName"
+                >
+                    <option [ngValue]="null">-</option>
+                    <option *ngFor="let g of groups$ | async" [ngValue]="g.name">
+                        {{g.name}}
+                    </option>
+                </select>
+            </div>
+            <div *ngIf="!hiddenFields.includes(HideableSubjectField.NAME)" class="form-group">
+                <label class="form-control-label"
+                       [translate]="'managementPortalApp.subject.personName'"
+                       for="field_person_name">
+                </label>
+                <input type="text" class="form-control" name="personName" id="field_person_name" [(ngModel)]="subject.personName"/>
+            </div>
 
-        <div *ngIf="!(siteSettings$ | async).hiddenSubjectFields.includes('group')" class="form-group">
-            <label class="form-control-label"
-                for="field_group"
-                [translate]="'managementPortalApp.subject.group'">
-            </label>
-            <select class="form-control"
-                id="field_group"
-                name="group"
-                [(ngModel)]="groupName"
-            >
-                <option [ngValue]="null">-</option>
-                <option *ngFor="let g of groups$ | async" [ngValue]="g.name">
-                    {{g.name}}
-                </option>
-            </select>
-        </div>
-        <div *ngIf="!(siteSettings$ | async).hiddenSubjectFields.includes('person_name')" class="form-group">
-            <label class="form-control-label"
-                   [translate]="'managementPortalApp.subject.personName'"
-                   for="field_person_name">
-            </label>
-            <input type="text" class="form-control" name="personName" id="field_person_name" [(ngModel)]="subject.personName"/>
-        </div>
-
-        <div *ngIf="!(siteSettings$ | async).hiddenSubjectFields.includes('date_of_birth')" class="form-group">
-            <label class="form-control-label"
-                   [translate]="'managementPortalApp.subject.dateOfBirth'"
-                   for="field_date_of_birth">
-            </label>
-            <div class="input-group">
-                <input [(ngModel)]="dateOfBirth"
-                       id="field_date_of_birth"
-                       class="form-control"
-                       name="dateOfBirth"
-                       placeholder="yyyy-mm-dd"
-                       ngbDatepicker
-                       #dateOfBirthDatePicker="ngbDatepicker">
-                <div class="input-group-append">
-                    <button class="btn btn-outline-secondary calendar" (click)="dateOfBirthDatePicker.toggle()" type="button"></button>
+            <div *ngIf="!hiddenFields.includes(HideableSubjectField.DATEOFBIRTH)" class="form-group">
+                <label class="form-control-label"
+                       [translate]="'managementPortalApp.subject.dateOfBirth'"
+                       for="field_date_of_birth">
+                </label>
+                <div class="input-group">
+                    <input [(ngModel)]="dateOfBirth"
+                           id="field_date_of_birth"
+                           class="form-control"
+                           name="dateOfBirth"
+                           placeholder="yyyy-mm-dd"
+                           ngbDatepicker
+                           #dateOfBirthDatePicker="ngbDatepicker">
+                    <div class="input-group-append">
+                        <button class="btn btn-outline-secondary calendar" (click)="dateOfBirthDatePicker.toggle()" type="button"></button>
+                    </div>
                 </div>
             </div>
         </div>
-
         <jhi-dictionary-mapper [attributes]='subject.attributes' [options]="options" [eventPrefix]='attributeComponentEventPrefix'></jhi-dictionary-mapper>
     </div>
     <div class="modal-footer">
