+++ src/main/webapp/app/shared/subject/subject.model.ts
@@ -45,7 +45,3 @@ export const enum SubjectStatus {
     'DISCONTINUED',
     'INVALID'
 }
-
-export interface SiteSettings {
-    hiddenSubjectFields: string[]
-}
